+++ docarray/array/doc_vec/doc_vec.py
@@ -18,7 +18,7 @@ from typing import (
 
 from pydantic import BaseConfig, parse_obj_as
 
-from docarray.array.abstract_array import AnyDocArray
+from docarray.array.any_array import AnyDocArray
 from docarray.array.doc_list.doc_list import DocList
 from docarray.array.doc_vec.column_storage import ColumnStorage, ColumnStorageView
 from docarray.array.doc_vec.list_advance_indexing import ListAdvancedIndexing
